requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-extensions==3.0.9
flake8==3.8.3
freeze==3.0
gunicorn==20.0.4
mccabe==0.6.1
Pillow==8.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
